docs: Custom Element
Developed this custom element to utilize in all views and index files needed. For a cleaner code file, and to render a nicer UI.

Element has enabled the UI to only display the links and actions pertaining to the page a user is on, rather then CakePHP's default side-nav which renders links and actions in a random pattern.